+++ packages/@n8n/benchmark/src/test-execution/app-metrics-poller.ts
@@ -1,11 +1,15 @@
 /**
  * Polls the /metrics endpoint from an n8n instance to collect application metrics
  * during benchmark test runs.
+ *
+ * A Poller can be started and stopped once. After it's stopped, it cannot be restarted.
+ * Instead, a new poller instance should be created.
  */
 export class AppMetricsPoller {
 	private intervalId: NodeJS.Timeout | undefined = undefined;
 	private metricsData: string[] = [];
 	private isRunning = false;
+	private isStopped = false;
 
 	constructor(
 		private readonly metricsUrl: string,
@@ -17,14 +21,10 @@ export class AppMetricsPoller {
 	 */
 	start() {
 		if (this.isRunning) {
-			console.warn('Metrics poller is already running');
-			return;
+			throw new Error('Metrics poller is already running');
 		}
-
-		// In case the old poller is somehow still running, make sure we stop it.
-		if (this.intervalId) {
-			clearInterval(this.intervalId);
-			this.intervalId = undefined;
+		if (this.isStopped) {
+			throw new Error('Metrics poller has been stopped and cannot be restarted');
 		}
 
 		this.isRunning = true;
@@ -48,6 +48,7 @@ export class AppMetricsPoller {
 			this.intervalId = undefined;
 		}
 		this.isRunning = false;
+		this.isStopped = true;
 	}
 
 	/**
@@ -76,68 +77,5 @@ export class AppMetricsPoller {
 				`Error polling metrics: ${error instanceof Error ? error.message : String(error)}`,
 			);
 		}
-	}
-}
-
-/**
- * Parses Prometheus metrics text format and extracts time series data
- */
-export class PrometheusMetricsParser {
-	/**
-	 * Extracts all values for a specific metric from collected metrics data
-	 */
-	static extractMetricValues(metricsData: string[], metricName: string): number[] {
-		const values: number[] = [];
-
-		for (const metricsText of metricsData) {
-			const lines = metricsText.split('\n');
-
-			for (const line of lines) {
-				// Skip comments and empty lines
-				if (line.startsWith('#') || line.trim() === '') {
-					continue;
-				}
-
-				// Parse metric line: metric_name{labels} value timestamp
-				// For simplicity, we'll just check if the line starts with the metric name
-				if (line.startsWith(metricName)) {
-					const parts = line.split(/\s+/);
-					if (parts.length >= 2) {
-						const value = parseFloat(parts[parts.length - 1]);
-						if (!isNaN(value)) {
-							values.push(value);
-						}
-					}
-				}
-			}
-		}
-
-		return values;
-	}
-
-	/**
-	 * Calculates statistics for a metric from collected data
-	 */
-	static calculateMetricStats(
-		metricsData: string[],
-		metricName: string,
-	): { max: number; avg: number; min: number; count: number } | null {
-		const values = this.extractMetricValues(metricsData, metricName);
-
-		if (values.length === 0) {
-			return null;
-		}
-
-		const max = Math.max(...values);
-		const min = Math.min(...values);
-		const sum = values.reduce((a, b) => a + b, 0);
-		const avg = sum / values.length;
-
-		return {
-			max,
-			min,
-			avg,
-			count: values.length,
-		};
 	}
 }
